# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * AdvocaciaHub API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
  message?: string;
}

export interface SuccessResponse {
  success: boolean;
  message?: string;
}

export interface Specialty {
  id: number;
  name: string;
  description?: string;
  parentId?: number | null;
  slug: string;
}

export interface LawyerSummary {
  id: number;
  fullName: string;
  oabNumber: string;
  oabState: string;
  city: string;
  state: string;
  bio: string;
  yearsExperience: number;
  hourlyRateMin?: number | null;
  hourlyRateMax?: number | null;
  isVerified: boolean;
  avgRating: number;
  totalReviews: number;
  photoUrl?: string | null;
  specialties: Specialty[];
}

export interface OfficeSummary {
  id: number;
  tradeName: string;
  city: string;
  state: string;
  lawyerCount: number;
  logoUrl?: string | null;
}

export type ReviewModerationStatus =
  (typeof ReviewModerationStatus)[keyof typeof ReviewModerationStatus];

export const ReviewModerationStatus = {
  pending: "pending",
  approved: "approved",
  rejected: "rejected",
} as const;

export interface Review {
  id: number;
  lawyerId: number;
  rating: number;
  comment: string;
  caseType: string;
  caseOutcome?: string | null;
  moderationStatus: ReviewModerationStatus;
  createdAt: string;
}

export type LawyerDetail = LawyerSummary & {
  phone?: string | null;
  website?: string | null;
  languages: string[];
  office?: OfficeSummary | null;
  recentReviews: Review[];
};

export interface LawyerListResponse {
  data: LawyerSummary[];
  total: number;
  page: number;
  limit: number;
  totalPages: number;
}

export type CreateReviewInputCaseType =
  (typeof CreateReviewInputCaseType)[keyof typeof CreateReviewInputCaseType];

export const CreateReviewInputCaseType = {
  trabalhista: "trabalhista",
  civil: "civil",
  criminal: "criminal",
  familia: "familia",
  tributario: "tributario",
  empresarial: "empresarial",
  consumidor: "consumidor",
  previdenciario: "previdenciario",
  outro: "outro",
} as const;

export type CreateReviewInputCaseOutcome =
  | (typeof CreateReviewInputCaseOutcome)[keyof typeof CreateReviewInputCaseOutcome]
  | null;

export const CreateReviewInputCaseOutcome = {
  vencido: "vencido",
  perdido: "perdido",
  acordo: "acordo",
  em_andamento: "em_andamento",
} as const;

export interface CreateReviewInput {
  lawyerId: number;
  /**
   * @minimum 1
   * @maximum 5
   */
  rating: number;
  /** @minLength 20 */
  comment: string;
  caseType: CreateReviewInputCaseType;
  caseOutcome?: CreateReviewInputCaseOutcome;
  clientName: string;
}

export interface BlogPost {
  id: number;
  title: string;
  excerpt: string;
  tags: string[];
  publishedAt: string;
  featuredImageUrl?: string | null;
  authorName: string;
  authorOab: string;
  readTimeMinutes: number;
}

export type BlogPostDetail = BlogPost & {
  content: string;
};

export interface BlogPostListResponse {
  data: BlogPost[];
  total: number;
  page: number;
  totalPages: number;
}

export type BannerPosition =
  (typeof BannerPosition)[keyof typeof BannerPosition];

export const BannerPosition = {
  header: "header",
  sidebar: "sidebar",
  "blog-between": "blog-between",
} as const;

export interface Banner {
  id: number;
  title: string;
  imageUrl: string;
  targetUrl: string;
  position: BannerPosition;
  isActive: boolean;
}

export interface ContactMessageInput {
  lawyerId: number;
  clientName: string;
  clientEmail: string;
  clientPhone?: string | null;
  message: string;
  caseType: string;
}

export type ListLawyersParams = {
  specialty?: string;
  city?: string;
  state?: string;
  minRating?: number;
  search?: string;
  page?: number;
  limit?: number;
};

export type ListBlogPostsParams = {
  tag?: string;
  page?: number;
  limit?: number;
};

export type ListBannersParams = {
  position?: ListBannersPosition;
};

export type ListBannersPosition =
  (typeof ListBannersPosition)[keyof typeof ListBannersPosition];

export const ListBannersPosition = {
  header: "header",
  sidebar: "sidebar",
  "blog-between": "blog-between",
} as const;