# lib/api-zod/src/generated/types/banner.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * AdvocaciaHub API
 * OpenAPI spec version: 0.1.0
 */
import type { BannerPosition } from "./bannerPosition";

export interface Banner {
  id: number;
  title: string;
  imageUrl: string;
  targetUrl: string;
  position: BannerPosition;
  isActive: boolean;
}