# lib/api-zod/src/generated/types/blogPost.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * AdvocaciaHub API
 * OpenAPI spec version: 0.1.0
 */

export interface BlogPost {
  id: number;
  title: string;
  excerpt: string;
  tags: string[];
  publishedAt: Date;
  featuredImageUrl?: string | null;
  authorName: string;
  authorOab: string;
  readTimeMinutes: number;
}