# lib/api-zod/src/generated/types/blogPostListResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * AdvocaciaHub API
 * OpenAPI spec version: 0.1.0
 */
import type { BlogPost } from "./blogPost";

export interface BlogPostListResponse {
  data: BlogPost[];
  total: number;
  page: number;
  totalPages: number;
}