# lib/api-zod/src/generated/types/contactMessageInput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * AdvocaciaHub API
 * OpenAPI spec version: 0.1.0
 */

export interface ContactMessageInput {
  lawyerId: number;
  clientName: string;
  clientEmail: string;
  clientPhone?: string | null;
  message: string;
  caseType: string;
}