# lib/api-zod/src/generated/types/createReviewInput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * AdvocaciaHub API
 * OpenAPI spec version: 0.1.0
 */
import type { CreateReviewInputCaseOutcome } from "./createReviewInputCaseOutcome";
import type { CreateReviewInputCaseType } from "./createReviewInputCaseType";

export interface CreateReviewInput {
  lawyerId: number;
  /**
   * @minimum 1
   * @maximum 5
   */
  rating: number;
  /** @minLength 20 */
  comment: string;
  caseType: CreateReviewInputCaseType;
  caseOutcome?: CreateReviewInputCaseOutcome;
  clientName: string;
}