# lib/api-zod/src/generated/types/lawyerDetail.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * AdvocaciaHub API
 * OpenAPI spec version: 0.1.0
 */
import type { LawyerSummary } from "./lawyerSummary";
import type { OfficeSummary } from "./officeSummary";
import type { Review } from "./review";

export type LawyerDetail = LawyerSummary & {
  phone?: string | null;
  website?: string | null;
  languages: string[];
  office?: OfficeSummary | null;
  recentReviews: Review[];
};