# lib/api-zod/src/generated/types/officeSummary.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * AdvocaciaHub API
 * OpenAPI spec version: 0.1.0
 */

export interface OfficeSummary {
  id: number;
  tradeName: string;
  city: string;
  state: string;
  lawyerCount: number;
  logoUrl?: string | null;
}